requirements.txt
Django==5.0.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
